+++ src/bridges/ethereum-darwiniaDVM/Ethereum2DarwiniaDVMRecord.tsx
@@ -1,3 +1,3 @@
-export function Ethereum2DarwiniaDVMRecord() {
-  return <span>Ethereum2DarwiniaDVMRecord</span>;
-}
+import { Ethereum2DarwiniaRecord } from '../ethereum-darwinia';
+
+export const Ethereum2DarwiniaDVMRecord = Ethereum2DarwiniaRecord;
