+++ src/bridges/substrate-dvm/Dvm2SubstrateRecord.tsx
@@ -1,3 +1,3 @@
-export function Dvm2SubstrateRecord() {
-  return <span>Dvm2SubstrateRecord</span>;
-}
+import { Substrate2DVMRecord } from './Substrate2DVMRecord';
+
+export const DVM2SubstrateRecord = Substrate2DVMRecord;
