+++ src/bridges/substrate-dvm/index.ts
@@ -1,2 +1,4 @@
 export * from './DVM2Substrate';
 export * from './Substrate2DVM';
+export * from './Substrate2DVMRecord';
+export * from './DVM2SubstrateRecord';
