--- src/components/record/ethereum-darwinia/index.ts
@@ -1,2 +1,0 @@
-export * from './Darwinia2Ethereum';
-export * from './Ethereum2Darwinia';
