--- src/components/record/ethereum-dvm/index.ts
@@ -1,4 +1,0 @@
-import { Darwinia2EthereumRecord, Ethereum2DarwiniaRecord } from '../ethereum-darwinia';
-
-export const Ethereum2DVMRecord = Ethereum2DarwiniaRecord;
-export const DVM2EthereumRecord = Darwinia2EthereumRecord;
