--- src/components/record/substrate-dvm/index.ts
@@ -1,1 +1,0 @@
-export * from './SubstrateDVM';
