--- src/components/record/substrate-substrateDVM/index.ts
@@ -1,1 +1,0 @@
-export * from './SubstrateSubstrateDVM';
